requirements.txt
h5py==2.7.0
Keras==2.0.6
numpy==1.13.1
tensorflow==1.2.1
Pillow==4.2.1
scikit-learn==0.19.0